CharMetrics: use short int instead of int. This halves the memory required for the font metrics cache.
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@16129 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/frontends/qt4/GuiFontMetrics.h
@@ -31,9 +31,9 @@ size_t const MaxCharType = 65536;
 
 struct CharMetrics
 {
-	int width;
-	int ascent;
-	int descent;
+	short int width;
+	short int ascent;
+	short int descent;
 };
